Decision management

Opens a Shiny app that shows a visual diff of each modified file.

Usage

import_review(
  source_path = "R/",
  output_path = get_target_dir(),
  background = getOption("autoimport_background", FALSE)
)

Source

inspired by testthat::snapshot_review()

Arguments

source_path

path to the original R files

output_path

path to the updated R files

background

whether to run the app in a background process. Default to getOption("autoimport_background", FALSE).

Value

nothing if background==FALSE, the (callr::process) object if background==TRUE

Warning

Beware that using background=TRUE can bloat your system with multiple R session!
You should probably kill the process when you are done:

p=import_review(background=TRUE)
p$kill()